refactor: Move Place to domain package
- 도메인 패키지를 추가
